Add Gate::authorize() to all Methods in Panel/PageController.php

+++ app/Http/Controllers/Panel/PageController.php
@@ -7,22 +7,29 @@ use App\Http\Requests\Panel\Page\CreatePageRequest;
 use App\Http\Requests\Panel\Page\UpdatePageRequest;
 use App\Models\Page;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class PageController extends Controller
 {
     public function index()
     {
+        Gate::authorize('view-pages');
+
         $pages = Page::orderByDesc('id')->paginate(5);
         return view('panel.pages.index',compact('pages'));
     }
 
     public function create()
     {
+        Gate::authorize('create-page');
+
         return view('panel.pages.create');
     }
 
     public function store(CreatePageRequest $request)
     {
+        Gate::authorize('store-page');
+
         $data = $request->validated();
 
         $file = $request->file('banner');
@@ -48,11 +55,15 @@ class PageController extends Controller
 
     public function edit(Page $page)
     {
+        Gate::authorize('edit-pages');
+
         return view('panel.pages.edit',compact('page'));
     }
 
     public function isApproved(Request $request, Page $page)
     {
+        Gate::authorize('is-approved-pages');
+
         $page->update([
             'is_approved' => ! $page->is_approved
         ]);
@@ -63,6 +74,8 @@ class PageController extends Controller
 
     public function update(UpdatePageRequest $request, Page $page)
     {
+        Gate::authorize('update-pages');
+
         $data = $request->validated();
 
         if ($request->hasFile('banner')){
@@ -82,6 +95,8 @@ class PageController extends Controller
 
     public function destroy(Page $page, Request $request)
     {
+        Gate::authorize('delete-pages');
+
         $page->delete();
         $request->session()->flash('status','صفحه تکی مورد نظر با موفقیت حذف شد !');
         return back();
